+++ src/core/shapes/Sphere.java
@@ -8,11 +8,11 @@ import core.world.Ray;
 public class Sphere implements Shape
 {
     private Vector3 pos;
-    private float radius;
+    private double radius;
     public Material mat;
-    public float distance;
+    public double distance;
 
-    public Sphere(float x, float y, float z, float r, Color color)
+    public Sphere(double x, double y, double z, double r, Color color)
     {
         this.pos = new Vector3(x, y, z);
         this.mat = new Material(color);
@@ -33,32 +33,32 @@ public class Sphere implements Shape
         Vector3 origin = ray.orig;
         Vector3 direction = ray.dir;
         Vector3 OC = Vector3.sub(pos, origin);
-        float tca = Vector3.dot(OC, direction);
+        double tca = Vector3.dot(OC, direction);
         //System.out.println(tca);
         if (tca < 0)
         {
             return false;
         }
-        float d2 = Vector3.dot(OC, OC) - (tca * tca);
+        double d2 = Vector3.dot(OC, OC) - (tca * tca);
         if (d2 > radius * radius)
         {
             return false;
         }
 
-        float thc = (float) Math.sqrt(radius * radius - d2);
-        float t0 = tca - thc;
-        float t1 = tca + thc;
+        double thc = Math.sqrt(radius * radius - d2);
+        double t0 = tca - thc;
+        double t1 = tca + thc;
         if (t0 > t1)
         {
-            float temp = t0;
+            double temp = t0;
             t0 = t1;
             t1 = temp;
         }
 
         Vector3 OC2 = Vector3.sub(origin, pos);
-        float a = Vector3.dot(direction, direction);
-        float b = 2 * Vector3.dot(OC2, direction);
-        float c = Vector3.dot(OC2, OC2) - (radius * radius);
+        double a = Vector3.dot(direction, direction);
+        double b = 2 * Vector3.dot(OC2, direction);
+        double c = Vector3.dot(OC2, OC2) - (radius * radius);
         if (!solveQuadratic(a, b, c, t0, t1))
         {
             return false;
@@ -78,32 +78,32 @@ public class Sphere implements Shape
         return true;
     }
 
-    public boolean solveQuadratic(float a, float b, float c, float x0, float x1)
+    public boolean solveQuadratic(double a, double b, double c, double x0, double x1)
     {
-        //float distanceCenter = (b*b) - (4 * c);
-        float discr = b * b - 4 * a * c;
+        //double distanceCenter = (b*b) - (4 * c);
+        double discr = b * b - 4 * a * c;
         if (discr < 0)
         {
             return false;
         } else if (discr == 0)
         {
-            x0 = x1 = (float) (-0.5 * b / a);
+            x0 = x1 = (double) (-0.5 * b / a);
         } else
         {
-            float q = 0;
+            double q = 0;
             if (b > 0)
             {
-                q = (float) (-0.5 * (b + (float) Math.sqrt(discr)));
+                q = (-0.5 * (b + Math.sqrt(discr)));
             } else
             {
-                q = (float) (-0.5 * (b - (float) Math.sqrt(discr)));
+                q = (-0.5 * (b - Math.sqrt(discr)));
             }
             x0 = q / a;
             x1 = c / q;
         }
         if (x0 > x1)
         {
-            float temp = x0;
+            double temp = x0;
             x0 = x1;
             x1 = temp;
         }
@@ -111,7 +111,7 @@ public class Sphere implements Shape
     }
 
     @Override
-    public float getDepth()
+    public double getDepth()
     {
         return this.distance;
     }
