+++ src/core/shapes/Triangle.java
@@ -9,8 +9,8 @@ import core.math.Vector3;
 public class Triangle
 {
     private Vector3 pos;
-    private float height;
-    private float width;
+    private double height;
+    private double width;
     private Color color;
 
     public Triangle()
